Move data from rules to actions
